# api.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: IssueTracker
  description: "This is the Open API documentation for the REST API of our beloved application **IssueTracker** deployed at <https://issuetrackerv1.herokuapp.com/>. <br>All operations are executable. Operations that make changes to the domain require authentication. In this case, you must **Authorize** your request by providing the token value you got when you created the tweet."
servers:
  - url: 'https://issuetrackerv1.herokuapp.com/'
paths:
...
  /issues/new:
    post:
      tags:
        - Issues
      summary: Adds a new Issues to the IssueTracker
      operationId: new_issue
      requestBody:
        description: Provide the Subject, Description, Assigned, Type, Severity, Priority Status and Watched of the new Issue
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                  example: 'Subject'
                description:
                  type: string
                  example: 'Description'
                assign:
                  type: string
                  example: 'Assign'
                severity:
                  type: string
                  example: 'Wishlist'
                priority:
                  type: string
                  example: 'High'
                issue_type:
                  type: string
                  example: 'Bug'
                status:
                  type: string
                  example: 'New'
                watcher:
                  type: array
                  items:
                    type: string
                  example: ['Gabriel del Valle']
      security:
        - api_key: []

      responses:
        200:
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error403'
components:
  schemas:
    Issue:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        subject:
          type: string
          example: 'Subject'
        description:
          type: string
          example: 'Description'
        assign:
          type: string
          example: 'Assign'
        created_at:
          type: string
          format: date-time
          example: 'Wed, 12 Apr 2023 13:52:06.671832000 UTC +00:00'
        updated_at:
          type: string
          format: date-time
          example: 'Wed, 12 Apr 2023 13:52:06.671832000 UTC +00:00'
        severity:
          type: string
          enum: ['Wishlist', 'Minor', 'Normal', 'Important', 'Critical']
          example: 'Wishlist'
        priority:
          type: string
          enum: ['Low', 'Normal', 'High']
          example: 'High'
        issue_type:
          type: string
          enum: ['Bug', 'Question', 'Enhancement']
          example: 'Bug'
        blocked:
          type: boolean
          example: true
        deadline:
          type: string
          format: date-time
          example: '05/05/2023'
        watcher:
          type: array
          items:
            type: string
          example: ['Gabriel del Valle']
        watcher_ids:
          type: array
          items:
            type: integer
            example: [1]
        status:
          type: string
          enum: ['New', 'In Progress', 'Ready For Test', 'Postponed', 'Closed', 'Information Needed', 'Rejected']
          example: 'New'
        created_by:
          type: string
          example: 'Gabriel del Valle'
      required:
        - id
        - created_at
        - severity
        - priority
        - issue_type
        - status
    ErrorResult:
      type: object
      properties:
        message:
          type: string
          example: 'Content is too long (maximum is 280 characters)'
      required:
        - message
  examples:
...
    error401:
      value:
        message: "You provided no token"
    error403:
      value:
        message: "You provided an invalid token"
...
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header